requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.11.7
email-validator
supabase==2.24.0
python-dotenv==1.2.1
